GILBERT HAS ARRIVED
oh no

+++ Wizard Arena/Assets/Scripts/ui/HealthBar.cs
@@ -5,8 +5,8 @@ using UnityEngine.UI;
 
 public class HealthBar : MonoBehaviour
 {
-	public int maxHealth;
-	public int currentHealth;
+	public float maxHealth;
+	public float currentHealth;
 
 	private RectTransform rt;
 
@@ -15,17 +15,16 @@ public class HealthBar : MonoBehaviour
         rt = gameObject.GetComponent<RectTransform>();
 	}
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
 	{
 		Debug.Log("max hp set");
 		maxHealth = health;
 		SetHealth(health);
 	}
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
 	{
 		currentHealth = health;
 		if (rt) rt.localScale = new Vector3((5.0f * currentHealth / maxHealth), rt.localScale.y, rt.localScale.z);
 	}
-
 }
